Fix test and line check failures
