+++ frontend/src/app/student/(authenticated)/dashboard/page.tsx
@@ -1,41 +1,16 @@
 "use client"
 
 import { useEffect, useMemo, useState } from "react"
-import Link from "next/link"
 import { CircleAlert } from "lucide-react"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import { Badge } from "@/components/ui/badge"
-import { Button } from "@/components/ui/button"
 import { StudentRecentResultsCard } from "@/components/student/student-recent-results-card"
 import { StudentScheduleCalendar } from "@/components/student/student-schedule-calendar"
+import { StudentUpcomingExamsCard } from "@/components/student/student-upcoming-exams-card"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
 import { examResults, exams as legacyExams, type Exam } from "@/lib/mock-data"
+import { getLocalDateString, getSecondsUntil } from "@/lib/student-exam-time"
 import { useStudentSession } from "@/hooks/use-student-session"
 import { getStudentExams } from "@/lib/student-exams"
-
-function getSecondsUntil(date: string, time: string) {
-  const examDate = new Date(`${date}T${time}:00`)
-  const now = new Date()
-  const diff = Math.floor((examDate.getTime() - now.getTime()) / 1000)
-  return diff > 0 ? diff : 0
-}
-
-function getLocalDateString() {
-  const now = new Date()
-  const year = now.getFullYear()
-  const month = String(now.getMonth() + 1).padStart(2, "0")
-  const day = String(now.getDate()).padStart(2, "0")
-  return `${year}-${month}-${day}`
-}
-
-function formatCountdown(seconds: number) {
-  const hours = Math.floor(seconds / 3600)
-  const minutes = Math.floor((seconds % 3600) / 60)
-  const secs = seconds % 60
-  if (hours > 0) return `${hours}h ${minutes}m ${secs}s`
-  if (minutes > 0) return `${minutes}m ${secs}s`
-  return `${secs}s`
-}
 
 export default function StudentDashboard() {
   const { studentClass, studentId, studentName } = useStudentSession()
@@ -112,47 +87,12 @@ export default function StudentDashboard() {
         </Alert>
       ) : null}
 
-      {/* Notifications / Upcoming Exams Alert */}
-      {upcomingExams.length > 0 && (
-        <Card className="panel-surface border-primary rounded-[1.5rem]">
-          <CardHeader className="pb-2">
-            <CardTitle className="text-base flex items-center gap-2">
-              Upcoming Exams
-              <Badge>{upcomingExams.length}</Badge>
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="space-y-2">
-              {upcomingExams.map(exam => {
-                const schedule = exam.scheduledClasses.find(sc => sc.classId === studentClass)
-                const isTodayExam = schedule?.date === today
-                const countdown = todayCountdowns[exam.id] ?? 0
-                return (
-                  <div
-                    key={exam.id}
-                    className={`elevated-surface flex items-center justify-between rounded-xl border p-3 soft-divider ${isTodayExam ? 'border-primary ring-1 ring-primary/40' : ''}`}
-                  >
-                    <div>
-                      <div className="font-medium">{exam.title}</div>
-                      <div className="secondary-text text-sm">
-                        {schedule?.date} at {schedule?.time} ({exam.duration} min)
-                      </div>
-                      {isTodayExam ? (
-                        <div className="mt-2 inline-flex items-center rounded-full bg-primary px-3 py-1 text-sm font-semibold text-primary-foreground">
-                          {countdown === 0 ? "Ready to start now" : `Starts in ${formatCountdown(countdown)}`}
-                        </div>
-                      ) : null}
-                    </div>
-                    <Link href={`/student/exams/${exam.id}`}>
-                      <Button size="sm" variant="outline">View</Button>
-                    </Link>
-                  </div>
-                )
-              })}
-            </div>
-          </CardContent>
-        </Card>
-      )}
+      <StudentUpcomingExamsCard
+        exams={upcomingExams}
+        studentClass={studentClass}
+        today={today}
+        todayCountdowns={todayCountdowns}
+      />
 
       {/* Quick Stats */}
       <div className="grid grid-cols-1 md:grid-cols-4 gap-4">
