+++ frontend/src/lib/student-exam-time.ts
@@ -1,0 +1,37 @@
+'use client'
+
+export function getSecondsUntil(date: string, time: string) {
+  const examDate = new Date(`${date}T${time}:00`)
+  const now = new Date()
+  const diff = Math.floor((examDate.getTime() - now.getTime()) / 1000)
+  return diff > 0 ? diff : 0
+}
+
+export function getLocalDateString() {
+  const now = new Date()
+  const year = now.getFullYear()
+  const month = String(now.getMonth() + 1).padStart(2, '0')
+  const day = String(now.getDate()).padStart(2, '0')
+  return `${year}-${month}-${day}`
+}
+
+export function formatCountdown(seconds: number) {
+  const hours = Math.floor(seconds / 3600)
+  const minutes = Math.floor((seconds % 3600) / 60)
+  const secs = seconds % 60
+  if (hours > 0) return `${hours}h ${minutes}m ${secs}s`
+  if (minutes > 0) return `${minutes}m ${secs}s`
+  return `${secs}s`
+}
+
+export function formatCountdownParts(seconds: number) {
+  const hours = Math.floor(seconds / 3600)
+  const minutes = Math.floor((seconds % 3600) / 60)
+  const secs = seconds % 60
+
+  return {
+    hours: hours.toString().padStart(2, '0'),
+    minutes: minutes.toString().padStart(2, '0'),
+    seconds: secs.toString().padStart(2, '0'),
+  }
+}
